temp

+++ AutoBar/AutoBar/Models/Order.cs
@@ -11,7 +11,7 @@ namespace AutoBar.Models
         public bool OrderStatus { get; set; }
         public bool HasReward { get; set; }
         public double TotalPrice { get; set; }
-        public int PointsEarned { get; set; }
+        public double PointsEarned { get; set; }
         public DateTime OpenedOn { get; set; }
         public DateTime ClosedOn { get; set; }
         public string CustomerName { get; set; }
